Split theme into multiple parts
