+++ lib/core/theme/theme.tailor.dart
@@ -2,7 +2,7 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint, unused_element, unnecessary_cast
 
-part of 'app_theme.dart';
+part of 'theme.dart';
 
 // **************************************************************************
 // TailorAnnotationsGenerator
