+++ lib/core/theme/theme_extensions.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+part of 'theme.dart';
 
 extension ThemeExtensions on BuildContext {
   TextTheme get textTheme => Theme.of(this).textTheme;
