+++ lib/features/home/settings_screen/settings_screen.dart
@@ -4,8 +4,7 @@ import 'package:der_die_das/core/db/settings/enums/number_questions.dart';
 import 'package:der_die_das/core/db/settings/state/settings_state.dart';
 import 'package:der_die_das/core/extensions/list_widget_extensions.dart';
 import 'package:der_die_das/core/l10n/l10n_extension.dart';
-import 'package:der_die_das/core/theme/app_theme.dart';
-import 'package:der_die_das/core/theme/theme_extensions.dart';
+import 'package:der_die_das/core/theme/theme.dart';
 import 'package:der_die_das/core/ui/common/basic_button.dart';
 import 'package:der_die_das/core/ui/common/basic_material_close_button.dart';
 import 'package:der_die_das/features/home/settings_screen/ui/language_button.dart';
